solution/fastapi_app/requirements.txt
annotated-types==0.6.0
anyio==4.3.0
asgiref==3.8.1
attrs==23.1.0
backoff==2.2.1
beautifulsoup4==4.12.0
certifi==2023.7.22
charset-normalizer==3.3.1
click==8.1.7
decorator==4.4.2
Deprecated==1.2.14
distlib==0.3.6
emoji==1.7.0
exceptiongroup==1.2.0
fastapi==0.108.0
filelock==3.10.7
gensim==3.8.3
googleapis-common-protos==1.63.0
grpcio==1.62.1
h11==0.14.0
httpcore==1.0.5
httpx==0.26.0
idna==3.4
imageio==2.34.0
imageio-ffmpeg==0.4.9
importlib-metadata==6.11.0
joblib==1.3.2
moviepy==1.0.3
nltk==3.8.1
numpy==1.24.4
opentelemetry-api==1.22.0
opentelemetry-distro==0.43b0
opentelemetry-exporter-otlp==1.22.0
opentelemetry-exporter-otlp-proto-common==1.22.0
opentelemetry-exporter-otlp-proto-grpc==1.22.0
opentelemetry-exporter-otlp-proto-http==1.22.0
opentelemetry-instrumentation==0.43b0
opentelemetry-instrumentation-asgi==0.43b0
opentelemetry-instrumentation-fastapi==0.43b0
opentelemetry-instrumentation-logging==0.43b0
opentelemetry-proto==1.22.0
opentelemetry-sdk==1.22.0
opentelemetry-semantic-conventions==0.43b0
opentelemetry-util-http==0.43b0
outcome==1.3.0.post0
pillow==10.3.0
platformdirs==3.2.0
proglog==0.1.10
prometheus-client==0.19.0
protobuf==4.25.3
pydantic==2.6.4
pydantic_core==2.16.3
pydub==0.25.1
pyopencl==2022.3.1
PySocks==1.7.1
pytools==2022.1.14
regex==2023.12.25
requests==2.31.0
scikit-learn==1.3.2
scipy==1.10.0
selenium==4.14.0
six==1.16.0
smart-open==7.0.4
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.5
SpeechRecognition==3.10.1
starlette==0.32.0.post1
text2emotion==0.0.5
threadpoolctl==3.4.0
tqdm==4.66.2
trio==0.22.2
trio-websocket==0.11.1
typing_extensions==4.10.0
urllib3==2.0.7
uvicorn==0.25.0
virtualenv==20.21.0
wrapt==1.16.0
wsproto==1.2.0
zipp==3.18.1
